Managed to make both DG and DeepONet work.
Changed JSON keys for DG to start with dg_ (both in DG JSONs as well as deeponet JSONs)
Added deeponet to the requirements file as an install from the git repository
Checking out the correct commit for DG when including it (so that it works with the write results functionality required by DeepONet)

Removed absorption override and reflection coefficient from the JSON files.
Added a plot_dg_results function to plot the uncorrected receiver results.

TODO: fully decouple deeponet from the DGinterface code. Currently, the DG file checks for a JSON entry to see if the function has been called from a DeepONet job. Preferably, DG should be fully decoupled from DeepONet.